AM-862 - Role Assignment Batch Job - Moved the Code to DeleteExpiredRecords
