+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/RoleAssignmentBatchApplication.java
@@ -6,9 +6,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.context.ApplicationContext;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import uk.gov.hmcts.reform.roleassignmentbatch.task.FetchExpiredRecords;
 
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class})
 @SuppressWarnings("HideUtilityClassConstructor")
